refactor gamma core table into split
so we can basically reuse our elasticsearch module

+++ modules/Shop/Jobs/Gamma/DeactivateProduct.php
@@ -3,6 +3,7 @@
 use App\Jobs\Job;
 use Illuminate\Contracts\Bus\SelfHandling;
 use Modules\Account\Account;
+use Modules\Shop\Gamma\ProductCategorySelection;
 use Modules\Shop\Gamma\ProductSelection;
 use Modules\Shop\Product\Category;
 use Modules\Shop\Product\Product;
@@ -30,7 +31,6 @@ class DeactivateProduct extends Job implements SelfHandling
             'account_id'  => $this->account->id,
             'product_id'  => $this->product->id,
             'brand_id'    => $this->product->brand_id,
-            'category_id' => $this->category->id,
         ];
 
         $record = $selection->newQuery()->where($payload)->first();
@@ -38,8 +38,46 @@ class DeactivateProduct extends Job implements SelfHandling
         if(!$record)
         {
             $record = $selection->create($payload);
+
+            $category = $this->attachCategory($record);
+        }
+        else{
+            $category_id = $this->category->id;
+
+            $record->load(['categories' => function($query) use ($category_id){
+                $query->where('category_id', $category_id);
+                $query->withTrashed();
+            }]);
+
+            $category = $record->categories->first();
+
+            if(!$category)
+            {
+                $category = $this->attachCategory($record);
+            }
         }
 
-        $record->delete();
+        //set them as deactivated.
+        $category->delete();
+
+        //only delete the base record when no more active categories are left.
+        if($record->categories()->count() == 0)
+        {
+            $record->delete();
+        }
+    }
+
+    /**
+     * @param $record
+     *
+     * @return ProductCategorySelection
+     */
+    protected function attachCategory($record)
+    {
+        $category = new ProductCategorySelection(['category_id' => $this->category->id]);
+
+        $record->categories()->save($category);
+
+        return $category;
     }
 }
